moved the generate_secret files to pkg/util and added the webhook condition in secret.go
