+++ pkg/util/generate_secret.go
@@ -1,4 +1,4 @@
-package utils
+package util
 
 import "crypto/rand"
