+++ pkg/util/generate_secret_test.go
@@ -1,4 +1,4 @@
-package utils
+package util
 
 import (
 	"testing"
